checkpoint 5pm debug
