+++ src/UI/corelated_stocks.py
@@ -7,8 +7,10 @@ from dotenv import load_dotenv
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..')))
 
 # Import the necessary agents and tasks
-from src.Agents.Analysis.stock_analysis_agent_correlated  import StockAnalysisAgents
-from src.Agents.Analysis.stock_analysis_task_correlated import StockAnalysisTasks
+#from src.Agents.Analysis.stock_analysis_agent_correlated  import StockAnalysisAgents
+#from src.Agents.Analysis.stock_analysis_task_correlated import StockAnalysisTasks
+from src.Agents.Correlation_Agents.correlation_agent import CorrelationAgent
+from src.Agents.Correlation_Agents.investment_decision_agent import InvestmentDecisionAgent
 
 load_dotenv()
 
@@ -18,21 +20,24 @@ class StockCorrelationCrew:
         self.stock2 = stock2
 
     def run(self):
-        agents = StockAnalysisAgents()
-        tasks = StockAnalysisTasks()
+        #agents = StockAnalysisAgents()
+        #tasks = StockAnalysisTasks()
 
-        # Initialize agents
-        correlation_agent = agents.correlation_analyst()
-        investment_advisor_agent = agents.investment_advisor()
+        # Initialize agents        
+        correlation_agent = CorrelationAgent()
+        investment_decision_agent = InvestmentDecisionAgent()
+        #investment_advisor_agent = agents.investment_advisor()
 
         # Initialize tasks
-        correlation_task = tasks.calculate_correlation(correlation_agent, self.stock1, self.stock2)
-        investment_task = tasks.investment_decision(investment_advisor_agent, self.stock1, self.stock2)
+        correlation_task = correlation_agent.calculate_correlation
+        investment_decision_task = investment_decision_agent.investment_decision
+        #correlation_task = tasks.calculate_correlation(correlation_agent, self.stock1, self.stock2)
+        #investment_task = tasks.investment_decision(investment_advisor_agent, self.stock1, self.stock2)
 
         # Use Crew to coordinate agents and tasks
         crew = Crew(
-            agents=[correlation_agent, investment_advisor_agent],
-            tasks=[correlation_task, investment_task],
+            agents=[correlation_agent, investment_decision_agent],
+            tasks=[correlation_task, investment_decision_task],
             verbose=True
         )
